mega patch to remove almost all compiler warnings on MacOS X where the compiler is by default in pedantic mode (LHCb also like to use this option). The following issues have been fixed: - removal of unused arguments - comparison between signed and unsigned integers - not calling of base class copy ctor in copy ctor's To be done, the TGeo classes where we get still many hundred warnings of the above nature. List forwarded to Andrei.
git-svn-id: http://root.cern.ch/svn/root/trunk@5717 27541ba8-7e3a-0410-8455-c3a389f83636

+++ clib/src/Getline.c
@@ -1,4 +1,4 @@
-/* @(#)root/clib:$Name:  $:$Id: Getline.c,v 1.11 2002/08/07 15:34:21 rdm Exp $ */
+/* @(#)root/clib:$Name:  $:$Id: Getline.c,v 1.12 2002/08/07 17:14:47 rdm Exp $ */
 /* Author: */
 
 /*
@@ -1362,8 +1362,8 @@ Gl_histadd(char *buf)
         len = strlen(buf);
         if (strchr(p, '\n'))    /* previously line already has NL stripped */
             len--;
-        if (prev == 0 || strlen(prev) != len ||
-                            strncmp(prev, buf, len) != 0) {
+        if (prev == 0 || (int)strlen(prev) != len ||
+            strncmp(prev, buf, len) != 0) {
             hist_buf[hist_last] = hist_save(buf);
             prev = hist_buf[hist_last];
             hist_last = (hist_last + 1) % HIST_SIZE;
